Move state helpers to stateHelpers.ts
