+++ services/app-api/src/testHelpers/index.ts
@@ -8,9 +8,9 @@ export { must } from './errorHelpers'
 
 export {
     createInsertContractData,
-    getStateRecord,
     createContractData,
     createContractRevision,
     createDraftContractData,
-    getProgramsFromState,
 } from './contractAndRates/contractHelpers'
+
+export { getProgramsFromState, getStateRecord } from './stateHelpers'
